+++ src/os2datascanner/engine2/pipeline/explorer.py
@@ -145,7 +145,8 @@ def message_received(  # noqa: CCR001
         yield messages.ProblemMessage(
                 scan_tag=message.scan_tag, source=message.source,
                 handle=getattr(message.source, "handle", None),
-                message=exception_message)
+                message=exception_message,
+                account_uuid=message.account_uuid)
         log.warning(
                 "finished unsuccessfully",
                 handle_count=handle_count, source_count=source_count,
@@ -171,7 +172,7 @@ def message_received(  # noqa: CCR001
                 objects_reported_individually=True,
                 sources_reported_individually=True,
                 message=exception_message,
-                status_is_error=exception_message != "")
+                status_is_error=exception_message != "" and message.account_uuid is None)
 
 
 def message_received_raw(body, channel, source_manager):